+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsEventParams.kt
@@ -15,18 +15,24 @@ import androidx.annotation.RestrictTo
  * @property endpoint The endpoint being called.
  * @property experiment Currently a ShopperInsights module specific event that indicates
  * the experiment, as a JSON string, that the merchant sent to the us.
- * @property paymentMethodsDisplayed A ShopperInsights module specific event that indicates the
- * order of payment methods displayed to the shopper by the merchant.
+ * @property shopperSessionId The Shopper Insights customer session ID created by a merchant's
+ * server SDK or graphQL integration.
+ * @property buttonType buttonType Represents the tapped button type.
+ * @property buttonOrder The order or ranking in which payment buttons appear.
+ * @property pageType The page or view that a button is displayed on.
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class AnalyticsEventParams @JvmOverloads constructor(
-    var payPalContextId: String? = null,
-    var linkType: String? = null,
-    var isVaultRequest: Boolean = false,
-    var startTime: Long? = null,
-    var endTime: Long? = null,
-    var endpoint: String? = null,
+    val payPalContextId: String? = null,
+    val linkType: String? = null,
+    val isVaultRequest: Boolean = false,
+    val startTime: Long? = null,
+    val endTime: Long? = null,
+    val endpoint: String? = null,
     val experiment: String? = null,
-    val paymentMethodsDisplayed: List<String> = emptyList(),
-    val appSwitchUrl: String? = null
+    val appSwitchUrl: String? = null,
+    val shopperSessionId: String? = null,
+    val buttonType: String? = null,
+    val buttonOrder: String? = null,
+    val pageType: String? = null
 )