+++ Demo/src/main/java/com/braintreepayments/demo/PayPalFragment.java
@@ -60,6 +60,7 @@ public class PayPalFragment extends BaseFragment {
         TextInputEditText buyerPhoneNationalNumberEditText = view.findViewById(R.id.buyer_phone_national_number_edit_text);
         Button billingAgreementButton = view.findViewById(R.id.paypal_billing_agreement_button);
         Button singlePaymentButton = view.findViewById(R.id.paypal_single_payment_button);
+
         Switch payPalErrorHandlingSwitch = view.findViewById(R.id.paypal_edit_error_request_toggle);
         TextInputEditText vaultIdEditText = view.findViewById(R.id.paypal_edit_vault_id_field);
         TextInputEditText riskCorrelationIdEditText = view.findViewById(
@@ -68,12 +69,15 @@ public class PayPalFragment extends BaseFragment {
 
         Button editVaultButton = view.findViewById(R.id.paypal_edit_vault_button);
 
+        Switch contactInformationSwitch = view.findViewById(R.id.contact_info_switch);
+
         singlePaymentButton.setOnClickListener(v -> {
             launchPayPal(
                 false,
                 buyerEmailEditText.getText().toString(),
                 buyerPhoneCountryCodeEditText.getText().toString(),
-                buyerPhoneNationalNumberEditText.getText().toString()
+                buyerPhoneNationalNumberEditText.getText().toString(),
+                contactInformationSwitch.isChecked()
             );
         });
         billingAgreementButton.setOnClickListener(v -> {
@@ -88,7 +92,8 @@ public class PayPalFragment extends BaseFragment {
                 true,
                 buyerEmailEditText.getText().toString(),
                 buyerPhoneCountryCodeEditText.getText().toString(),
-                buyerPhoneNationalNumberEditText.getText().toString()
+                buyerPhoneNationalNumberEditText.getText().toString(),
+                false
             );
         });
 
@@ -224,7 +229,8 @@ public class PayPalFragment extends BaseFragment {
         boolean isBillingAgreement,
         String buyerEmailAddress,
         String buyerPhoneCountryCode,
-        String buyerPhoneNationalNumber
+        String buyerPhoneNationalNumber,
+        Boolean isContactInformationEnabled
     ) {
         FragmentActivity activity = getActivity();
         activity.setProgressBarIndeterminateVisibility(true);
@@ -236,10 +242,10 @@ public class PayPalFragment extends BaseFragment {
                 if (dataCollectorResult instanceof DataCollectorResult.Success) {
                     deviceData = ((DataCollectorResult.Success) dataCollectorResult).getDeviceData();
                 }
-                launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber);
+                launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber, isContactInformationEnabled);
             });
         } else {
-            launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber);
+            launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber, isContactInformationEnabled);
         }
     }
 
@@ -249,7 +255,8 @@ public class PayPalFragment extends BaseFragment {
         String amount,
         String buyerEmailAddress,
         String buyerPhoneCountryCode,
-        String buyerPhoneNationalNumber
+        String buyerPhoneNationalNumber,
+        Boolean isContactInformationEnabled
     ) {
         PayPalRequest payPalRequest;
         if (isBillingAgreement) {
@@ -257,7 +264,8 @@ public class PayPalFragment extends BaseFragment {
                 activity,
                 buyerEmailAddress,
                 buyerPhoneCountryCode,
-                buyerPhoneNationalNumber
+                buyerPhoneNationalNumber,
+                    "fake-session-id"
             );
         } else {
             payPalRequest = createPayPalCheckoutRequest(
@@ -265,7 +273,9 @@ public class PayPalFragment extends BaseFragment {
                 amount,
                 buyerEmailAddress,
                 buyerPhoneCountryCode,
-                buyerPhoneNationalNumber
+                buyerPhoneNationalNumber,
+                isContactInformationEnabled,
+                    null
             );
         }
         payPalClient.createPaymentAuthRequest(requireContext(), payPalRequest, (paymentAuthRequest) -> {