+++ Demo/src/main/java/com/braintreepayments/demo/ShopperInsightsFragment.kt
@@ -18,6 +18,11 @@ import com.braintreepayments.api.paypal.PayPalPaymentAuthRequest
 import com.braintreepayments.api.paypal.PayPalPaymentAuthResult
 import com.braintreepayments.api.paypal.PayPalPendingRequest
 import com.braintreepayments.api.paypal.PayPalResult
+import com.braintreepayments.api.shopperinsights.ButtonOrder
+import com.braintreepayments.api.shopperinsights.ButtonType
+import com.braintreepayments.api.shopperinsights.ExperimentType
+import com.braintreepayments.api.shopperinsights.PageType
+import com.braintreepayments.api.shopperinsights.PresentmentDetails
 import com.braintreepayments.api.shopperinsights.ShopperInsightsBuyerPhone
 import com.braintreepayments.api.shopperinsights.ShopperInsightsClient
 import com.braintreepayments.api.shopperinsights.ShopperInsightsRequest
@@ -60,17 +65,21 @@ class ShopperInsightsFragment : BaseFragment() {
     private lateinit var venmoStartedPendingRequest: VenmoPendingRequest.Started
     private lateinit var paypalStartedPendingRequest: PayPalPendingRequest.Started
 
+    private var shopperSessionId: String = "test-shopper-session-id"
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        shopperInsightsClient = ShopperInsightsClient(requireContext(), authStringArg)
+        shopperInsightsClient = ShopperInsightsClient(requireContext(), authStringArg, shopperSessionId)
 
         venmoClient = VenmoClient(requireContext(), super.getAuthStringArg(), null)
         payPalClient = PayPalClient(
-            requireContext(), super.getAuthStringArg(),
-            Uri.parse("https://mobile-sdk-demo-site-838cead5d3ab.herokuapp.com/")
+            requireContext(),
+            super.getAuthStringArg(),
+            Uri.parse("https://mobile-sdk-demo-site-838cead5d3ab.herokuapp.com/braintree-payments"),
+            "com.braintreepayments.demo.braintree"
         )
 
         return inflater.inflate(R.layout.fragment_shopping_insights, container, false)
@@ -191,17 +200,25 @@ class ShopperInsightsFragment : BaseFragment() {
                 is ShopperInsightsResult.Success -> {
                     if (result.response.isPayPalRecommended) {
                         payPalVaultButton.isEnabled = true
-                        shopperInsightsClient.sendPayPalPresentedEvent(
-                            """{"exp_name":"PaymentReady","treatment_name":"control"}""",
-                            listOf("PayPal", "Venmo", "other")
+                        shopperInsightsClient.sendPresentedEvent(
+                            ButtonType.PAYPAL,
+                            PresentmentDetails(
+                                ExperimentType.TEST,
+                                ButtonOrder.FIRST,
+                                PageType.HOMEPAGE
+                            )
                         )
                     }
 
                     if (result.response.isVenmoRecommended) {
                         venmoButton.isEnabled = true
-                        shopperInsightsClient.sendVenmoPresentedEvent(
-                            """{"exp_name":"PaymentReady","treatment_name":"test"}""",
-                            listOf("Venmo", "PayPal", "other")
+                        shopperInsightsClient.sendPresentedEvent(
+                            ButtonType.VENMO,
+                            PresentmentDetails(
+                                ExperimentType.TEST,
+                                ButtonOrder.OTHER,
+                                PageType.HOMEPAGE
+                            )
                         )
                     }
 
@@ -221,7 +238,9 @@ class ShopperInsightsFragment : BaseFragment() {
     }
 
     private fun launchPayPalVault() {
-        shopperInsightsClient.sendPayPalSelectedEvent()
+        shopperInsightsClient.sendSelectedEvent(
+            ButtonType.PAYPAL
+        )
 
         payPalClient.createPaymentAuthRequest(
             requireContext(),
@@ -229,7 +248,9 @@ class ShopperInsightsFragment : BaseFragment() {
                 activity,
                 emailInput.editText?.text.toString(),
                 countryCodeInput.editText?.text.toString(),
-                nationalNumberInput.editText?.text.toString()
+                nationalNumberInput.editText?.text.toString(),
+                shopperSessionId
+
             )
         ) { authRequest ->
             when (authRequest) {
@@ -257,7 +278,9 @@ class ShopperInsightsFragment : BaseFragment() {
     }
 
     private fun launchVenmo() {
-        shopperInsightsClient.sendVenmoSelectedEvent()
+        shopperInsightsClient.sendSelectedEvent(
+            ButtonType.VENMO
+        )
 
         val venmoRequest = VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE)
         venmoRequest.profileId = null