+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalClient.kt
@@ -10,9 +10,9 @@ import com.braintreepayments.api.core.BraintreeException
 import com.braintreepayments.api.core.BraintreeRequestCodes
 import com.braintreepayments.api.core.Configuration
 import com.braintreepayments.api.core.ExperimentalBetaApi
+import com.braintreepayments.api.core.GetReturnLinkUseCase
 import com.braintreepayments.api.core.LinkType
 import com.braintreepayments.api.core.TokenizationKey
-import com.braintreepayments.api.core.GetReturnLinkUseCase
 import com.braintreepayments.api.core.MerchantRepository
 import com.braintreepayments.api.core.UserCanceledException
 import com.braintreepayments.api.paypal.PayPalPaymentIntent.Companion.fromString
@@ -38,7 +38,6 @@ class PayPalClient internal constructor(
     private val merchantRepository: MerchantRepository = MerchantRepository.instance,
     private val getReturnLinkUseCase: GetReturnLinkUseCase = GetReturnLinkUseCase(merchantRepository)
 ) {
-
     /**
      * Used for linking events from the client to server side request
      * In the PayPal flow this will be either an EC token or a Billing Agreement token
@@ -54,6 +53,11 @@ class PayPalClient internal constructor(
      * True if `tokenize()` was called with a Vault request object type
      */
     private var isVaultRequest = false
+
+    /**
+     * Used for sending Shopper Insights session ID provided by merchant to FPTI
+     */
+    private var shopperSessionId: String? = null
 
     /**
      * The final URL string used to open the app switch flow
@@ -93,11 +97,13 @@ class PayPalClient internal constructor(
      * @param payPalRequest a [PayPalRequest] used to customize the request.
      * @param callback      [PayPalPaymentAuthCallback]
      */
+    @OptIn(ExperimentalBetaApi::class)
     fun createPaymentAuthRequest(
         context: Context,
         payPalRequest: PayPalRequest,
         callback: PayPalPaymentAuthCallback
     ) {
+        shopperSessionId = payPalRequest.shopperSessionId
         isVaultRequest = payPalRequest is PayPalVaultRequest
 
         braintreeClient.sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_STARTED, analyticsParams)
@@ -537,7 +543,8 @@ class PayPalClient internal constructor(
             return AnalyticsEventParams(
                 payPalContextId = payPalContextId,
                 linkType = linkType?.stringValue,
-                isVaultRequest = isVaultRequest
+                isVaultRequest = isVaultRequest,
+                shopperSessionId = shopperSessionId
             )
         }