+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalLauncherUnitTest.kt
@@ -1,6 +1,7 @@
 package com.braintreepayments.api.paypal
 
 import android.content.Intent
+import android.net.Uri
 import androidx.activity.ComponentActivity
 import com.braintreepayments.api.BrowserSwitchClient
 import com.braintreepayments.api.BrowserSwitchException
@@ -11,6 +12,9 @@ import com.braintreepayments.api.core.ExperimentalBetaApi
 import com.braintreepayments.api.paypal.vaultedit.PayPalVaultEditAuthResult
 import com.braintreepayments.api.paypal.vaultedit.PayPalVaultEditPendingRequest
 import com.braintreepayments.api.core.AnalyticsClient
+import com.braintreepayments.api.core.AnalyticsEventParams
+import com.braintreepayments.api.core.GetReturnLinkUseCase
+import com.braintreepayments.api.core.MerchantRepository
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
@@ -32,13 +36,26 @@ class PayPalLauncherUnitTest {
     private val options: BrowserSwitchOptions = mockk(relaxed = true)
     private val pendingRequestString = "pending_request_string"
     private val analyticsClient: AnalyticsClient = mockk(relaxed = true)
+    private val merchantRepository = mockk<MerchantRepository>(relaxed = true)
+    private val getReturnLinkUseCase = mockk<GetReturnLinkUseCase>()
+    private val returnUrl = "https://return.url"
+    private val deepLinkScheme = "deepLinkScheme"
 
     private lateinit var sut: PayPalLauncher
 
     @Before
     fun setup() {
         every { paymentAuthRequestParams.browserSwitchOptions } returns options
-        sut = PayPalLauncher(browserSwitchClient, lazy { analyticsClient })
+
+        val appSwitchReturnUrl = Uri.parse(returnUrl)
+        every { getReturnLinkUseCase() } returns GetReturnLinkUseCase.ReturnLinkResult.AppLink(
+            appSwitchReturnUrl
+        )
+        sut = PayPalLauncher(
+            browserSwitchClient,
+            merchantRepository,
+            getReturnLinkUseCase,
+            lazy { analyticsClient })
     }
 
     @Test
@@ -103,7 +120,48 @@ class PayPalLauncherUnitTest {
             PayPalPendingRequest.Started(pendingRequestString),
             intent
         )
-        verify { analyticsClient.sendEvent(PayPalAnalytics.HANDLE_RETURN_STARTED) }
+        verify {
+            analyticsClient.sendEvent(
+                PayPalAnalytics.HANDLE_RETURN_STARTED,
+                AnalyticsEventParams(appSwitchUrl = returnUrl)
+            )
+        }
+    }
+
+    @Test
+    @Throws(JSONException::class)
+    fun `handleReturnToApp with deeplinkScheme sends handle started event with deeplink scheme`() {
+        every { getReturnLinkUseCase() } returns GetReturnLinkUseCase.ReturnLinkResult.DeepLink(
+            deepLinkScheme
+        )
+        sut.handleReturnToApp(
+            PayPalPendingRequest.Started(pendingRequestString),
+            intent
+        )
+        verify {
+            analyticsClient.sendEvent(
+                PayPalAnalytics.HANDLE_RETURN_STARTED,
+                AnalyticsEventParams(appSwitchUrl = deepLinkScheme)
+            )
+        }
+    }
+
+    @Test
+    @Throws(JSONException::class)
+    fun `handleReturnToApp with ReturnLinkResult Failure sends handle started event with null appSwitchUrl`() {
+        every { getReturnLinkUseCase() } returns GetReturnLinkUseCase.ReturnLinkResult.Failure(
+            Exception("handle return start failed")
+        )
+        sut.handleReturnToApp(
+            PayPalPendingRequest.Started(pendingRequestString),
+            intent
+        )
+        verify {
+            analyticsClient.sendEvent(
+                PayPalAnalytics.HANDLE_RETURN_STARTED,
+                AnalyticsEventParams()
+            )
+        }
     }
 
     @Test