+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Braintree Android SDK Release Notes
 
+## 5.6.0 (2025-02-05)
+
+* ShopperInsights (BETA)
+  * Add `isPayPalAppInstalled` and `isVenmoAppInstalled` methods
+  * Add `shopperSessionId` parameter to `ShopperInsightsClient`
+* PayPal
+  * Add `shopperSessionId` to `PayPalCheckoutRequest` and `PayPalVaultRequest`
+  * Replace `sendPayPalPresentedEvent()` and `sendVenmoPresentedEvent()` with `sendPresentedEvent()`
+  * Replace `sendPayPalSelectedEvent()` and `sendVenmoSelectedEvent()` with `sendSelectedEvent()`
+
+## 5.5.0 (2025-01-23)
+
+* PayPal
+  * Add `PayPalContactInformation` request object
+  * Add `PayPalCheckoutRequest.contactInformation` optional property
+  * Fix bug where `intent=order` was not being set as expected
+
+## 5.4.0 (2025-01-21)
+ 
+* PayPal
+  * Fix bug to ensure that `PayPalVaultRequest.userAuthenticationEmail` is not sent as an empty string
+  * Add `shippingCallbackUrl` to `PayPalCheckoutRequest`
+* ThreeDSecure
+  * Return error if no `dfReferenceId` is returned in the 3D Secure flow
+
 ## 5.3.0 (2024-12-11)
 
 * PayPal